moved pika parser to submodule
